refactor(cli): cli 目录三段式收口 — commands / lib / oclif
按 reviewer 进一步建议把 src/cli/ 内顶层杂乱的 utility / type / parse helper
分层成清晰的三段:

```
src/cli/
  index.ts                # 入口
  commands/               # 产品 CLI 命令树
  lib/                    # 跨命令共享 utility
  oclif/                  # oclif 框架 plumbing
```

3 个变化:

A. 删 src/cli/parse-strict.ts(dead code):
   PR #124 双 parser 拆解后 src/ 业务 0 处 import parseArgsStrictOrExit;
   unknown flag exit 2 现在由 oclif `failedFlagParsing` 直接处理。删 src 文件,
   同步删 cli-exit.test.ts 内自验 case + 调整 oclif-doctor.test.ts 注释。

B. types/cmd-flags.ts → lib/cmd-flags.ts:
   reviewer C2 建议把 cmd-flags 收敛到 commands/ 同级;实操发现 oclif 默认
   把 commands/ 目录里所有 .ts(含 underscore 前缀)都当 Command 加载,
   `commands/types.ts` / `commands/_types.ts` 都触发「command types not
   found」warning,污染 stderr 让 test fail。退而求其次落 lib/cmd-flags.ts
   仍在 CLI namespace,跟其它 utility 同居(reviewer C2 主要建议「保持
   CLI namespace、不要进 src/types」已满足)。

C. 8 个顶层 utility file → src/cli/lib/:
   - cli-exit / i18n / i18n-dict / parse-run-config / progress / run-tally /
     update-check 直接搬
   - _shared.ts → lib/shared.ts(underscore 前缀历史是 internal marker,
     搬进 lib/ 后 explicit directory 已表达)
   - 8 个 file 内部 cross-src import 加一层 `..`(跳 lib/)
   - 所有 importer(src/cli/index.ts / commands/* / oclif/i18n.ts /
     src/doctor/* / src/eval-workflows/* / test/*)import path 重写
   - `oclif/i18n.ts` 跟 `lib/i18n.ts` 职责不同(前者是 oclif Command static
     字段双语 sentinel + LangAwareHelp 渲染入口,后者是 runtime tCli /
     tBoth / getCliLang / parseLangFromArgv 基础设施),不合并(reviewer
     已确认强行扁平反而混)

CONTRIBUTING.md「CLI 走 oclif 框架」section 加三段式目录布局表 + 更新业务
住哪儿描述(跨命令 utility 进 src/cli/lib/);加 oclif/i18n vs lib/i18n 区分
说明。

yarn lint && build && build:docs:check && test 1498 全绿
(parse-strict 删 → -1 case)。CLI smoke + OMK_LANG env 正常。
