+++ src/cli/commands/doctor.ts
@@ -2,8 +2,8 @@ import { existsSync, statSync } from 'node:fs';
 import { dirname, join, resolve } from 'node:path';
 import { Args, Command, Flags } from '@oclif/core';
 import { bilingual, resolveLang } from '../oclif/i18n.js';
-import { CliExit } from '../cli-exit.js';
-import { tCli } from '../i18n.js';
+import { CliExit } from '../lib/cli-exit.js';
+import { tCli } from '../lib/i18n.js';
 import type { Sample } from '../../types/index.js';
 import type { DependencyRequirements } from '../../eval-core/dependency-checker.js';
