+++ src/cli/commands/eval/gold/compare.ts
@@ -1,9 +1,9 @@
 import { resolve } from 'node:path';
 import { Args, Command, Flags } from '@oclif/core';
 import { bilingual, resolveLang } from '../../../oclif/i18n.js';
-import { CliExit } from '../../../cli-exit.js';
-import { DEFAULT_REPORTS_DIR } from '../../../parse-run-config.js';
-import { requireEvaluationReport } from '../../../_shared.js';
+import { CliExit } from '../../../lib/cli-exit.js';
+import { DEFAULT_REPORTS_DIR } from '../../../lib/parse-run-config.js';
+import { requireEvaluationReport } from '../../../lib/shared.js';
 import type { ReportStore } from '../../../../types/index.js';
 
 export default class EvalGoldCompare extends Command {
