+++ src/cli/commands/eval/gold/validate.ts
@@ -1,6 +1,6 @@
 import { Args, Command, Flags } from '@oclif/core';
 import { bilingual, resolveLang } from '../../../oclif/i18n.js';
-import { CliExit } from '../../../cli-exit.js';
+import { CliExit } from '../../../lib/cli-exit.js';
 
 export default class EvalGoldValidate extends Command {
   static description = bilingual({
