+++ src/cli/lib/i18n-dict.ts
