+++ src/cli/lib/i18n.ts
@@ -1,4 +1,4 @@
-import type { Lang } from '../types/shared.js';
+import type { Lang } from '../../types/shared.js';
 import { CLI_DICT, type CliMessageKey } from './i18n-dict.js';
 
 export type CliLang = Lang;
