+++ src/cli/lib/progress.ts
