+++ src/cli/lib/update-check.ts
