+++ test/cli/cli-exit.test.ts
@@ -1,8 +1,7 @@
 import { describe, it, expect } from 'vitest';
-import { CliExit } from '../../src/cli/cli-exit.js';
+import { CliExit } from '../../src/cli/lib/cli-exit.js';
 import { runEvolve } from '../../src/cli/commands/evolve.js';
-import { parseArgsStrictOrExit } from '../../src/cli/parse-strict.js';
-import { parseJudgeModelsArgOrExit } from '../../src/cli/parse-run-config.js';
+import { parseJudgeModelsArgOrExit } from '../../src/cli/lib/parse-run-config.js';
 
 /**
  * CliExit 收口让业务 runX() 在单测里可以被 try/catch 捕获,不再 kill 整个
@@ -30,15 +29,6 @@ describe('CliExit dispatch', () => {
     ).then(() => null, (e: unknown) => e);
     expect(err).toBeInstanceOf(CliExit);
     expect((err as CliExit).code).toBe(1);
-  });
-
-  it('parseArgsStrictOrExit 收到未知 option 时 throw CliExit(2)', () => {
-    expect(() =>
-      parseArgsStrictOrExit({
-        args: ['--bogus-flag'],
-        options: { known: { type: 'string' } },
-      }),
-    ).toThrowError(CliExit);
   });
 
   // parseJudgeModelsArgOrExit 也要走 CliExit，否则 eval / evolve 子命令单测里
