+++ test/cli/parse-judge-models-arg.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from 'vitest';
 import assert from 'node:assert/strict';
-import { parseJudgeModelsArg } from '../../src/cli/parse-run-config.js';
+import { parseJudgeModelsArg } from '../../src/cli/lib/parse-run-config.js';
 
 /**
  * `parseJudgeModelsArg` 是 `--judge-models executor:model[,executor:model,...]`
